+++ packages/cspell-lib/src/Models/ValidationIssue.ts
@@ -1,0 +1,7 @@
+import type { ExtendedSuggestion } from './Suggestion';
+import type { ValidationResult } from './ValidationResult';
+
+export interface ValidationIssue extends ValidationResult {
+    suggestions?: string[];
+    suggestionsEx?: ExtendedSuggestion[];
+}